fix pull request

+++ project/src/main/java/com/example/project/controller/UserController.java
@@ -8,6 +8,7 @@ import com.example.project.model.User;
 import com.example.project.dto.response.UserCreateResponse;
 import com.example.project.dto.response.UserResponse;
 import com.example.project.repository.IRoleRepository;
+import com.example.project.repository.IUserRepository;
 import com.example.project.service.IUserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
@@ -23,9 +24,9 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
-import static java.net.HttpURLConnection.HTTP_NOT_FOUND;
-import static java.net.HttpURLConnection.HTTP_OK;
+import static java.net.HttpURLConnection.*;
 
 @RequiredArgsConstructor
 @RestController
@@ -33,6 +34,7 @@ import static java.net.HttpURLConnection.HTTP_OK;
 public class UserController {
     private final IUserService userService;
     private  final IRoleRepository iRoleRepository;
+    private final IUserRepository iUserRepository;
     private final MessageSource messageSource;
     // Get list user
     @GetMapping("") // http://localhost:3000/api/users?page=1&limit=5
@@ -47,33 +49,41 @@ public class UserController {
 
     // create user
     @PostMapping("")
-    public ResponseEntity<?> createUser(@RequestBody @Valid UserCreateRequestDTO request, BindingResult result) {
-        try {
-            if (result.hasErrors()) {
-                List<String> errMessage = result.getAllErrors()
-                        .stream()
-                        .map(fieldError -> fieldError.getDefaultMessage())
-                        .toList();
-                return ResponseEntity.badRequest().body(errMessage);
-            }
-            UserDTO newUser = userService.createUser(request);
-            var userModel = UserCreateResponse.builder()
-                    .firstName(newUser.getFirstName())
-                    .lastName(newUser.getLastName())
-                    .email(newUser.getEmail())
-                    .roleId(newUser.getRoleId())
-                    .address(newUser.getAddress())
-                    .phone(newUser.getPhone())
-                    .build();
-            return ResponseEntity.ok(userModel);
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
+    public ResponseEntity<ResponseMessage<UserDTO>>  createUser(@RequestBody UserCreateRequestDTO request) throws Exception {
+        if(iUserRepository.existsByEmail(request.getEmail())){
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(messageSource.getMessage("ERROR_EMAIL_ALREADY_EXISTS",
+                            null,
+                            LocaleContextHolder.getLocale()), messageSource.getMessage("RESOURCE_NOT_FOUND_CODE",
+                            null,
+                            LocaleContextHolder.getLocale())));
+        }
+        UserDTO addedUser = userService.createUser(request);
+        if (addedUser != null) {
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(messageSource.getMessage("ACCOUNT_REGISTER_SUCCESS",
+                            null,
+                            LocaleContextHolder.getLocale()), addedUser));
+        } else {
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(messageSource.getMessage("ACCOUNT_REGISTER_FAILED",
+                            null,
+                            LocaleContextHolder.getLocale()), messageSource.getMessage("HTTP_INTERNAL_SERVER_ERROR",
+                            null,
+                            LocaleContextHolder.getLocale())));
         }
     }
     // Get single user
     @GetMapping("/{id}")
     public ResponseEntity<?> getUser(@PathVariable Long id) {
         try {
+            Optional<User> accountDelete = iUserRepository.findById(id);
+            //Check if account has been deleted
+            if(accountDelete.get().isDeleteFlag())
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(new ResponseMessage<>(messageSource.getMessage("USER_IS_DELETED",
+                                null,
+                                LocaleContextHolder.getLocale()) , HTTP_OK));
             User user = userService.getUserByID(id);
             return ResponseEntity.ok(UserResponse.mapUser(user));
         } catch (Exception e) {
@@ -83,7 +93,15 @@ public class UserController {
     }
     //Update user
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateUser(@Valid @PathVariable Long id, @Valid @RequestBody UserDTO request) {
+    public ResponseEntity<ResponseMessage<UserDTO>> updateUser(@Valid @PathVariable Long id, @Valid @RequestBody UserDTO request) {
+        if(iUserRepository.existsByEmail(request.getEmail())){
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(messageSource.getMessage("ERROR_EMAIL_ALREADY_EXISTS",
+                            null,
+                            LocaleContextHolder.getLocale()), messageSource.getMessage("RESOURCE_NOT_FOUND_CODE",
+                            null,
+                            LocaleContextHolder.getLocale())));
+        }
         UserDTO accountUpdated = userService.updateUser(id, request);
         if (accountUpdated != null) {
             return ResponseEntity.status(HttpStatus.OK)
@@ -98,12 +116,19 @@ public class UserController {
                             null,
                             LocaleContextHolder.getLocale())));
         }
-
     }
     // Delete User
     @DeleteMapping("/{id}")
     public ResponseEntity<ResponseMessage<UserDTO>> deleteUser(@PathVariable Long id) {
         try {
+            Optional<User> accountDelete = iUserRepository.findById(id);
+            //Check if account has been deleted
+            if(accountDelete.get().isDeleteFlag())
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(new ResponseMessage<>(messageSource.getMessage("USER_IS_DELETED",
+                                null,
+                                LocaleContextHolder.getLocale()) , HTTP_OK));
+            //delete account
             userService.deleteUser(id);
             return ResponseEntity.status(HttpStatus.OK)
                     .body(new ResponseMessage(messageSource.getMessage("USER_DELETE_SUCCESS",
@@ -121,9 +146,9 @@ public class UserController {
         }
         catch (Exception e) {
             return ResponseEntity.status(HttpStatus.OK)
-                    .body(new ResponseMessage(messageSource.getMessage("USER_NOT_FOUND",
+                    .body(new ResponseMessage(messageSource.getMessage("ACCOUNT_DELETE_FAILD",
                             null,
-                            LocaleContextHolder.getLocale()), messageSource.getMessage("RESOURCE_NOT_FOUND_CODE",
+                            LocaleContextHolder.getLocale()), messageSource.getMessage("ACCOUNT_DELETE_FAILD",
                             null,
                             LocaleContextHolder.getLocale())));
         }
@@ -138,11 +163,14 @@ public class UserController {
             return ResponseEntity.status(HttpStatus.OK)
                     .body(new ResponseMessage(messageSource.getMessage("ROLES_GET_ALL_FAILED",
                             null,
-                            LocaleContextHolder.getLocale()), HTTP_NOT_FOUND));
+                            LocaleContextHolder.getLocale()), messageSource.getMessage("RESOURCE_NOT_FOUND_COD",
+                            null,
+                            LocaleContextHolder.getLocale())));
         }
         return ResponseEntity.status(HttpStatus.OK)
                 .body(new ResponseMessage(messageSource.getMessage("ROLES_GET_ALL_SUCCESS",
                         null,
                         LocaleContextHolder.getLocale()), listRole));
     }
+
 }
